Kicked lua signal names out in favour of vector
